requirements.txt
sarge==0.1.4
PyYAML==3.11
requests==2.7.0
